Refatora RegisteredIssuesManager.__init__
